+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.21.2] - 2024-06-15
+
+This version includes preliminaries for supporting
+`define-public-module`.
+
 ## [0.21.1] - 2024-06-14
 
 This version includes the logic to resolve module names for use in